Use renamed GridContents.js
